# clean-architecture/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Product API with authentication",
    "title": "Go Expert API Example",
    "termsOfService": "https://swagger.io/terms/",
    "contact": {
      "name": "Henrique Figueiredo",
      "url": "",
      "email": "henriquerfigueiredo@outlook.com"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/products": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "List Products",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "parameters": [
          {
            "type": "string",
            "description": "page number",
            "name": "page",
            "in": "query"
          },
          {
            "type": "string",
            "description": "limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "order type",
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/entity.Product"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Create Product",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "parameters": [
          {
            "description": "product request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateProductInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.CreateProductOutput"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Get Product",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Product ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entity.Product"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Update Product",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Product ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Product",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateProductInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entity.Product"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Delete Product",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Product ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "Create User",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "parameters": [
          {
            "description": "user request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateUserInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      }
    },
    "/users/auth": {
      "post": {
        "description": "Get a user JWT",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "parameters": [
          {
            "description": "user credentials",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.GetJWTInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.GetJWTOutput"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.CreateProductInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    },
    "dto.CreateProductOutput": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    },
    "dto.CreateUserInput": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "dto.GetJWTInput": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.GetJWTOutput": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
    "entity.Product": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}